requirements.txt
Django==4.0.3
django-bootstrap-v5==1.0.11
django-widget-tweaks==1.4.12
django-cities-light==3.9.0
kafka-python==2.0.2
Pillow==9.0.1
black==22.1.0